AuthServer: Revert spurious 2to3 change from "int" to "long"
The name "long" (as opposed to "short") here refers to the
length of the message stored in the variable, so it
shouldn't have been renamed when converting to Python 3.

It's unfortunate that the variable shared a name with a
built-in type, but now that "long" isn't a built-in any more,
it shouldn't be a problem.